requirements.txt
blessed==1.20.0
PyDictionary==2.0.1
PyDictionary==2.0.1
pyenchant==3.2.2
pyfiglet==1.0.2
spacy==3.6.1
termcolor==2.3.0
